refactor(feature-flagging): separate agent runtime and telemetry
